Default car sound improvements
Improves the awful, ugly-sounding default car sounds so the lower
quality mods aren't so bad. I added the nice sounding sounds and
modified the default soundscript a bit to use not just idle sounds but
also low, med, and high rpm sounds. This is tested with my copy of RoR
and works fine. Sounds used:
http://www.rigsofrods.com/repository/view/5289