:triangular_ruler: Removed broken "cablight" feature.
An Ogre::Light (type: point light) was being set up when 1st cinecam was added to a vehicle. Apparently the motivation was to illuminate the vehicle interior when driving at night and using cabin view. But there's a bug: the light is created at position 0,0,0 and never moved. From player feedback it appears the feature never worked right.