:video_game: Static collisions now truly unlimited
Intro: static collision objects on terrain are split into equal-size 2D "cells" and a hashtable is used to find the cell.

Original implementation of the hashtable had a restriction: each entry must contain collision elements of only one "cell". Other cells with same hash would get a different entry assigned. In practice, this caused the hashtable to become full and limit the number of collision elements on map. The author(s) are no longer around to consult.

The new implementation is simplified: the hashtable is now an ordinary hashtable and hash collisions are handled by linear search and matching by "Cell ID".