:video_game: Preselected terrain only effective on game start.
When user returns to menu and clicks [Singleplayer], he can choose a map like usual, but the value of "Preselected Terrain" RoR.cfg entry is preserved and can be changed in settings window.

To support this behavior, the GVar concept was extended: a 'Stored' value was added. This can be either a standalone value or an alias of 'Active' value. See updated commentary in Application.h

Fixes #1353