:video_game: 4th remake of GVars
The GVar concept was extended to accomodate values which need to change at runtime but retain the original configured value, for example FOV. See updated comments in 'Application.h' for details. This change was already done before for 'GVarStr', but this revision extends it to all other GVar types and cleans up the code.

USABILITY IMPROVEMENTS:
* Added GVar 'mp_join_on_startup' corresponding to old RoR.cfg entry "Network enabled". This enables it to be configured in Settings window (TODO: add a widget for it)
* Corrected setting of FOV in Settings window - the entered value will now propagate to RoR.cfg on exit.
* Editing GVars in multiplayer window unified with settings dialog - same widgets are used, there's no SAVE button.
* GVar log entries are now more intuitive.