:wrench: Greatly simplified SlideNodes
BEFORE: a 'RailGroup' was a hand-made linked list of 'Rail' classes. This required complex memory management logic and an utility class 'RailBuilder'. Further, when searching through the entire RailGroup, an ugly code was required to detect loop in 'Rail' instances and stop the search.

AFTER: a 'RailGroup' is a std::vector<RailSegment>, where 'RailSegment' is renamed and cleaned up 'Rail' - it still contains the linked-list-style pointers (good for skipping to neighbours since the RailGroup can be a closed loop) but the full search is simply done over the std::vector<>. Also, the RailBuilder utility and a lot of other cruft was removed.