:wrench: Eliminated `IHeightFinder`, terrain added to `App`
Removed interface 'IHeightFinder' and associated function 'TerrainGeometryManager::getHeightFinder()'. Functions 'getHeightAt()' and 'getNormalAt()' moved directly to 'TerrainManager'.

TerrainManager was removed from 'gEnv' and put to 'App' namespace as 'GetSimTerrain()/SetSimTerrain()'

Removed a lot of 'getHeightFinder()/getGeometryManager()' calls used only as "is terrain loaded?" check - this is a leftover from times when simulation was started before the terrain was actually available.