:wrench: :bulb: Architecture: player vehicle info migrated.
The 'current player vehicle' and 'previous player vehicle' states are now kept in RoRFrameListener (future name: SimController) instead of ActorManager.

RATIONALE: ActorManager manages physics which shouldn't need the info, and SimController manages gfx/game-logic which needs the info. Previously this was handled by `ChangedCurrentVehicle()` callback and a lot of "what is player's vehicle?" queries. You'll notice that this revision deletes a lot of `//TODO: This isn't right` comments.