:triangular_ruler: Goodbye `RoRFrameListener`
The class was renamed to `SimController` - a long planned change, it was already referred to as such in most of the code. NOTE: Filename was temporarily kept unchanged to avoid modifying CMakeLists.txt and rebuilding.

Also, access to it was unified to use `App::GetSimController()`  - there were a lot of leftover raw pointers (and associated getters/setters) from a short era when I wanted to remove all globals from the game. But it turns out there is always some global state (rendering device, sound device, controller devices, OS state etc...), so why bother?