:triangular_ruler: Project ActorClass: eliminated `rig_t`
Project ActorClass is a tidy-up and refactor operation of RoR's softbody objects, historically called 'truck' (initial RoR was heavy truck simulator), then 'Beam' (with a subclass of 'rig_t'). The future name will be 'Actor' because it best describes the nature of the softbody objects in RoR: it can be anything from a soda can to a complex machine.