:bug: :bulb: Fix crash after removing vehicle (+refactor)
BUG: Typo in file "BeamFactory.cpp" on line "1340" - "int m_simulated_actor" dummy value was 1 instead of -1

REFACTOR: The "simulated actor" variable was made local to ActorManager::UpdateActors() because it's not used anywhere else and the mechanism is all weird, I want to remove it. Note: there was also access in ActorManager::WakeUpAllActors() but I deleted the code because it didn't make sense - why alter behavior of AIRPLANE-s based on invoking "wake up all actors" from GUI menubar or script?