:construction: :bulb: Project AsyncScene: Introduced `GfxScene` object
The `SimController::UpdateGfxScene()` is not enough to run scene updates asynchronously in the future - some global simulation state needs to be buffered (such as 'current player vehicle') and the best way to code it is to add a new object encapsulating this logic.

See comments/code in 'DustManager.h' for more info.