:truck: Truck spawner now fully sequential again
In 2013 I tried to make the parser analyze and group everything so the spawner woudln't have to obey the order of elements in truck file. It was supposed to help creating Blender-like ingame editor. It was a bad idea and this commit fully restores the original sequential processing.

Almost all elements in truck file are now placed in the 'sequence' array and processed in that order. See struct ActorSpawnState - the variables are almost-exact copy of v0.4.0.7's parser. To compare, see https://github.com/only-a-ptr/rigs-of-rods/tree/retro-0407