separated character animations to upper/lower body.
Instead of single animation channel, there are now 2. This commit should make it obvious how the data flow thru.

The code which applies the animation to OGRE objects is now different - instead of looping through all anims and detecting which one we want to update, we kill all first and then re-enable what we want.
Note that the upper/lower categories are informal, any anymation may control any part of the body, or whole. All existing anims are submitted as lowerbody, but upperbody would work just as well.