Added hydro flag `j`, restored quirky `i` for backw. compat.
It turns out the `i` flag has a bug: it always makes the hydro invisible, but if it comes as first/single flag, then it also makes the hydro commanded by steering input. That means:
* `ri` will make the hydro respond only to airplane rudder, and invisible.
* `ir` will make the hydro respond to both airplane rudder and steering input, and invisible.

To resolve backward compatibility, I added a new flag `j`, which makes the hydro invisible but without the quirks of `i`.

The documentation https://docs.rigsofrods.org/vehicle-creation/fileformat-truck/#hydros needs to be updated:
* `` no flags given -> force `n` flag.
* `n` means Normal and makes the hydro driven by steering input.
and the above.