:video_game: Enabled character/actor collision in multiplayer.
This is an experimental spinoff from #3049. I felt inspired.
Since the collision works by "sticking" the character to the actor while in contact, I realized I could extend our existing driver-attachment logic to also handle this attachment.It's glitchy right now, partly because the networked cab offset is always in world coordinates, so it causes sliding when the vehicle turns.

Minor cleanup: removed the `enabled` param from `SetActorCoupling()` because it was dummy.