Fixed hooks not being thread safe.
*O*U*C*H* for the 5th time!

This is a follow up to the MSG_SIM_ACTOR_LINKING_REQUESTED commit which fixes an omission in hooks. Unlike ropes and ties which just add an inter-actor beam and are done, hooks continuously shorten the beam until a defined boundary is reached. So while ropes and ties did the inter-beaming on main thread as I intended, the hooks still did it on physics worker threads.

CODECHANGES:
- SimData.h - removed PREUNLOCK, unlocking is now synchronous.
- ActorForcesEuler.cpp - immediate removal of interbeam replaced by HOOK_UNLOCK request via message queue. Adding the interbeam was moved to `hookToggle()`
- Actor.cpp - `hookToggle()` - interbeam is now added synchronously.