:sparkles: Added asset pack support :package:
A new file type "*.assetpack" was added to modcache with all the usual treatment:
- the file is recorded to the cache file (/cache/cache.json),
- it's searchable and selectable via UI (`LT_AssetPack`)
- can have a thumbnail (filename format '{}-mini.*'' where {} is the assetpack file name without extension)
- The file itself can be empty, recognized entries are `assetpack_name` (display name), `assetpack_description` (for display in selector UI) and `assetpack_author` (same syntax as in rig-def (aka .truck) file).

Assetpacks are referenced by the file name (standard approach) and can be used for both terrains and actors:
- Terrains: in .terrn2 file, add section `[AssetPacks]` followed by list of filenames (with extension, each on separate line) ending with `=` token, same as in `[Objects/Scripts]` sections - this is a quirk of the game.
- Actors: in the .truck file, use section `assetpacks` followed by list of filenames (with extension, each on separate line)

Note that, unlike any other mod, assetpacks aren't loaded to their own resource group but rather to the resoruce group of the mod which requested them. Therefore try to name all resources in an unique way to avoid clashes.