Tuning: added :envelope: MSG_GUI_REFRESH_TUNING_MENU_REQUESTED
This removes the huge UI lag on first open of the tuning menu, which was caused by loading all addonparts to evaluate conflicts. But the lag isnt gone - it now happens upon spawning the actor. It could only be fully eliminated by caching the tweaked element IDs directly in `CacheEntry`, something I'll eventually do but not now.

Codechanges:
* CacheSystem: added dummy TuneupDef with addonpart data to CacheEntry.
* AddonPartFileFormat: added silent mode, using the new CacheEntry field.
* GameContext: also push MSG_SIM_REFRESH_TUNING_MENU_REQUESTED after spawning new actor and seating player there. Deleted duplicate code in SpawnActor()