Collision perf: first test collcab aabb with actor aabb.
When searching for colliding nodes in the pre-buffered `contactable_point_pool`, first check if the collcab AABB intersects with the 'collision partner' actor's AABB, only then search the actor's nodes for intersection.
To do this, we need to know where each actor starts and ends in the `contactable_point_pool`. For this purpose, `struct ColActor` was added.