:bulb: Added optional 'Bundle-qualified' syntax, i.e. "mybundle.zip:myactor.truck"
Problem: in many scenarios, the game select an actor by truckfile name alone:
 *   savegames
 *   preselected truck in RoR.cfg
 *   preloaded trucks on terrain (.tobj)
 *   Multiplayer spawn notifications

Solution: introduce an optional "Bundle-qualified" syntax, i.e. "mybundle.zip:myactor.truck" (in case of ZIP archive) or "mysubdir:myproject.truck" (in case of subdir) and use it in all the above cases. Because it's optional, it doesn't need a version bump in savegame format or RoRnet protocol. Note that 'bundle' is RoR jargon for a ZIP archive or directory under /mods.

Codechanges:
* CacheSystem: completely eliminated helpers `CheckResourceLoaded()` which searched across all resource groups, potentially introducing mixups. Extended `FindEntryByFilename()` to support the new Bundle-qualified syntax.
* TerrainObjectManager: don't check preselected trucks immediatelly, it will be done on spawn anyway.
* Savegame: always use the Bundle-qualified syntax.
* Actor: in multiplayer, always use the Bundle-qualified syntax when announcing spawn.