Fixed VisualStudio warning C4858: discarding return value
warning C4858: discarding return value: This member function returns a bool indicating whether the collection is empty and has no other effects. It is not useful to call this member function and discard the return value. This collection can be cleared by assigning an empty value to it.

using VisualStudio - Treat warnings level1 as errors