:sparkles: Add option to force obstruction filter when inside an actor
There is no simple way to tell whether a truck has a closed cabin or not. This option exists to force the obstruction filter inside an actor if desired.

Since the check if the listener is inside the player's coupled actor is based on the actor's AAB, the obstruction filter might get enabled if the listener is close to but not inside the actor.

Future work: it might be better to perform line of sight checks against the mesh of a vehicle, which could solve both problems.